修改文件浏览sdcard挂载状态的监听等

+++ app/src/main/java/com/askey/dvr/cdr7010/filemanagement/service/FileManagerService.java
@@ -7,7 +7,9 @@ import android.os.RemoteException;
 
 import com.askey.dvr.cdr7010.filemanagement.IFileManagerAidlInterface;
 import com.askey.dvr.cdr7010.filemanagement.controller.MediaScanner;
+import com.askey.dvr.cdr7010.filemanagement.util.Const;
 import com.askey.dvr.cdr7010.filemanagement.util.Logg;
+import com.askey.dvr.cdr7010.filemanagement.util.SdcardUtil;
 
 import java.util.List;
 
@@ -26,6 +28,7 @@ public class FileManagerService extends Service {
         Intent startIntent = new Intent(this, SdcardService.class);
         startService(startIntent);
 
+        Const.SDCARD_IS_EXIST = SdcardUtil.checkSdcardExist();
         return new MyBinder();
     }
 
@@ -34,31 +37,42 @@ public class FileManagerService extends Service {
         @Override
         public boolean openSdcard() throws RemoteException {
 
+            if(Const.SDCARD_IS_EXIST){
+
+            }
             return false;
         }
 
         @Override
         public boolean closeSdcard() throws RemoteException {
+            if(Const.SDCARD_IS_EXIST){
 
+            }
             return false;
         }
 
         @Override
         public List<String> getAllFilesByType(String type) throws RemoteException {
-
-            return MediaScanner.getAllFileList(type);
+            if(Const.SDCARD_IS_EXIST){
+                return MediaScanner.getAllFileList(type);
+            }
+            return null;
         }
 
         @Override
         public boolean deleteFile(String path) throws RemoteException {
-
-            return MediaScanner.deleteFile(path);
+            if(Const.SDCARD_IS_EXIST){
+                return MediaScanner.deleteFile(path);
+            }
+            return false;
         }
 
         @Override
         public boolean deleteFileByFolder(String path) throws RemoteException {
-
-            return MediaScanner.deleteDirectory(path);
+            if(Const.SDCARD_IS_EXIST){
+                return MediaScanner.deleteDirectory(path);
+            }
+            return false;
         }
 
     }
